requirements.txt
altgraph==0.17.4
blinker==1.8.2
packaging==24.1
pefile==2024.8.26
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyserial==3.5
pywin32-ctypes==0.2.3
setuptools==74.0.0
